# public/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a test task",
    "version": "1.0.0",
    "title": "School"
  },
  "host": "localhost:83",
  "basePath": "/api",
...
  "schemes": ["http"],
  "paths": {
    "/student": {

      "post": {
        "tags": ["student"],
        "summary": "Add a new student",
        "description": "",
        "operationId": "addStudent",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Student params",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Student"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created success",
            "schema":{
              "type":"integer",
              "format":"int64"
            }
          },
          "400": {"description": "Bad request"}
        }
      },

      "get": {
        "tags": ["student"],
        "summary": "Get list existing students",
        "description": "",
        "operationId": "getStudents",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Success",
            "schema":{
              "type":"array",
              "items":{
                "$ref": "#/definitions/Student"
              }
            }
          }
        }
      }
    },
    "/student/{id}": {

      "put": {
        "tags": ["student"],
        "summary": "Update an existing student",
        "description": "",
        "operationId": "updateStudent",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of student to update",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "New student params",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Student"
            }
          }
        ],
        "responses": {
          "200": {"description": "Updated success"},
          "404": {"description": "Student not found"}
        }
      },

      "get": {
        "tags": ["student"],
        "summary": "Get an existing student",
        "description": "",
        "operationId": "getStudent",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of student to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema":{
                "$ref": "#/definitions/Student"
            }
          },
          "404": {"description": "Student not found"}
        }
      },

      "delete": {
        "tags": ["student"],
        "summary": "Remove an existing student",
        "description": "",
        "operationId": "removeStudent",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of student to remove",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "204": {"description": "Removed success"},
          "404": {"description": "Student not found"}
        }
      }
    },
...
  },
  "definitions": {
    "Student": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "first_name": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        },
        "group": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
  }
}